+++ CHANGELOG.md
@@ -43,3 +43,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.4.0] - 2017-10-28
 ### Added
 - Integration with Robinhood auth token handling on api
+
+## [0.5.0] - 2017-10-30
+### Added
+- App authentication routing based on authentication states
+- User email address in header on login